+++ lib/devise/orm.rb
@@ -1,10 +1,22 @@
 module Devise
-  module OrmDirtyTracking # :nodoc:
-    def self.activerecord51?
-      defined?(ActiveRecord) && ActiveRecord.gem_version >= Gem::Version.new("5.1.x")
+  module Orm # :nodoc:
+    def self.active_record?(model)
+      defined?(ActiveRecord) && model < ActiveRecord::Base
     end
 
-    if activerecord51?
+    def self.active_record_51?(model)
+      active_record?(model) && ActiveRecord.gem_version >= Gem::Version.new("5.1.x")
+    end
+
+    def self.included(model)
+      if Devise::Orm.active_record_51?(model)
+        model.include DirtyTrackingNewMethods
+      else
+        model.include DirtyTrackingOldMethods
+      end
+    end
+
+    module DirtyTrackingNewMethods
       def devise_email_before_last_save
         email_before_last_save
       end
@@ -28,7 +40,9 @@ module Devise
       def devise_respond_to_and_will_save_change_to_attribute?(attribute)
         respond_to?("will_save_change_to_#{attribute}?") && send("will_save_change_to_#{attribute}?")
       end
-    else
+    end
+
+    module DirtyTrackingOldMethods
       def devise_email_before_last_save
         email_was
       end